CREATE/se crea la rama 12 para la entrega del proyecto final. se borran los console.log, y las funciones y componentes que no se utilizaban. se modifican ultimos estilos del responsive
