+++ README.md
@@ -1,11 +1,15 @@
 
-
-
 # *** 🌱 Ecommerce Insumos Vivero Meme 🪴***
 
-## Aplicación desarrollada con [React](https://es.reactjs.org/) como proyecto del curso de Coderhouse. La misma es una simulacion de eccomerce de insumos de un vivero. 
+## Aplicación desarrollada con [React](https://es.reactjs.org/) como proyecto del curso de Coderhouse. La misma es responsive y es una simulacion de eccomerce de insumos de un vivero. 
 <br>
 
+### 📍***Sigue estos comandos una vez hayas clonado el repositorio:***
+1- git init
+2- npm start (Runs the app in the development mode.
+Open http://localhost:3000 to view it in your browser.)
+
+<br>
 
 ### 📍***Este proyecto utiliza las siguientes dependencias:***
 📁 [React Router](https://reactrouter.com/)
@@ -13,6 +17,9 @@
 📁 [Bootstrap-React](https://react-bootstrap.netlify.app/)
 
 📁 [Fontawesome](https://fontawesome.com/v5/docs/web/use-with/react)
+
+📁 [Firebase](https://firebase.google.com/?hl=es-419&gclid=CjwKCAjw0a-SBhBkEiwApljU0rw1Ff-0X6-wLIs8axQBc2L1U2R_6pjtNoKfOJIo_qUj8HoYndSlqxoCyToQAvD_BwE&gclsrc=aw.ds)
+
 
 ### Hecho con amor por mi 💕
