+++ src/components/Checkout/Checkout.css
@@ -14,6 +14,15 @@
     margin-top: 20px;
 }
 
+.container-input{
+    margin-bottom: 10px;
+}
+
+.container-label{
+    font-style:italic;
+    font-weight: 600;
+    margin-top: 10px;
+}
 
 @media (max-width: 1000px){
     .form-container{
