+++ src/components/Error/Error.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import './Error.css'
 import { Link } from 'react-router-dom';
 import { Button } from 'react-bootstrap';
+
+
 const Error = () => {
     return (
         <div className="error-container">
