+++ src/components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import './ItemDetailContainer.css'
-// import useFetch from '../hooks/useFetch';
 import ItemDetail from '../ItemDetail/ItemDetail';
 import Loading from '../Loading/Loading';
 import { useParams } from 'react-router-dom';
@@ -10,8 +9,6 @@ import { db } from '../../utils/firebase';
 
 
 const ItemDetailContainer = ({setModalShow, id }) => {
-
-    // const url = 'https://60a6e6f9b970910017eb2862.mockapi.io/users/articles/'
 
     const {itemIdParams} = useParams();
     const fecthId = itemIdParams ? itemIdParams : id;
@@ -27,9 +24,6 @@ const ItemDetailContainer = ({setModalShow, id }) => {
             setModalShow(false)
         }
     }
-
-    // const urlId = `${url}${fecthId}`
-    // const data = useFetch(urlId)
 
     useEffect(() => {
         setLoading(true)
