+++ src/components/ItemListContainer/ItemListContainer.css
@@ -1,6 +1,6 @@
 .itemList-Container{
     padding: 2em;
-    margin-top: 20%;
+    margin-top: 25%;
 }
 
 .title-categories{
